refactor(architecture): use bloc pattern and add tests
