+++ lib/api/api_service.dart
@@ -41,4 +41,28 @@ class ApiService {
       throw Exception('Cannot get attendee');
     }
   }
+
+  Future<List<Attendee>> getVisitors() async {
+    final user =
+        User.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_USER)));
+    final accessToken = await storage.read(key: STORAGE_KEY_ACCESS_TOKEN);
+    final event =
+        Event.fromJson(jsonDecode(await storage.read(key: STORAGE_KEY_EVENT)));
+
+    final response = await http.get(
+        '${DotEnv().env[ENV_KEY_API_URL]}/${user.tenant}/events/${event.id}/visitors',
+        headers: {HttpHeaders.authorizationHeader: "Bearer $accessToken"});
+
+    if (response.statusCode == 200) {
+      final _rawVisitors = jsonDecode(response.body);
+      final _visitors = List<Attendee>();
+      for (var rawEvent in _rawVisitors) {
+        _visitors.add(Attendee.fromJson(rawEvent));
+      }
+      _visitors.sort((a, b) => a.firstName.compareTo(b.firstName));
+      return _visitors;
+    } else {
+      throw Exception('Failed to load visitors');
+    }
+  }
 }
