+++ lib/common/bloc.dart
@@ -1,0 +1,3 @@
+abstract class Bloc {
+  void dispose();
+}
