+++ lib/common/bloc_provider.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/widgets.dart';
+
+import 'bloc.dart';
+
+class BlocProvider<T extends Bloc> extends StatefulWidget {
+  final Widget child;
+  final T bloc;
+
+  const BlocProvider({Key key, @required this.bloc, @required this.child})
+      : super(key: key);
+
+  static T of<T extends Bloc>(BuildContext context) {
+    return context.findAncestorWidgetOfExactType<BlocProvider<T>>().bloc;
+  }
+
+  @override
+  State createState() => _BlocProviderState();
+}
+
+class _BlocProviderState extends State<BlocProvider> {
+  @override
+  Widget build(BuildContext context) => widget.child;
+
+  @override
+  void dispose() {
+    widget.bloc.dispose();
+    super.dispose();
+  }
+}
