+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:x_qrcode/event/events_screen.dart';
 import 'package:x_qrcode/home/home_screen.dart';
 import 'package:x_qrcode/visitor/visitor_screen.dart';
 import 'package:x_qrcode/visitor/consent_screen.dart';
-import 'package:x_qrcode/visitor/visitors_screen.dart';
+import 'package:x_qrcode/visitors/visitors_screen.dart';
 
 import 'attendee/attendees_screen.dart';
 import 'auth/login_screen.dart';
