+++ pubspec.yaml
@@ -9,18 +9,19 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^0.17.3+1
+  barcode_scan: ^2.0.1
   flutter_auth0: 0.5.0
-  flutter_secure_storage: 3.3.1+1
   flutter_dotenv: ^2.1.0
-  cupertino_icons: ^0.1.2
+  flutter_secure_storage: 3.3.1+1
+  flutter_sticky_header: ^0.4.2
+  flutter_svg: ^0.17.3+1
   http: ^0.12.0+4
   signature: ^3.0.0
-  barcode_scan: ^2.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  mockito: ^4.1.1
 
 flutter:
   uses-material-design: true
